合并了所有分支

+++ pinyougou-parent/pinyougou-service-buyer/src/main/java/cn.itcast.core.service/OrderPayServiceImpl.java
@@ -127,7 +127,7 @@ public class OrderPayServiceImpl implements OrderPayService {
                 orderVo.setOrderId(order.getOrderId());
                 orderVo.setOrderItemId(orderItem.getItemId());
                 orderVo.setBigPrice(goods.getPrice());
-                orderVo.setPrice(orderItem.getPrice());
+                orderVo.setPrice(orderItem.getPrice().doubleValue());
                 orderVo.setNum(orderItem.getNum());
                 orderVo.setTotalFee(orderItem.getTotalFee());
                 orderVo.setSourceType(order.getSourceType());
